==> 03_basic_data_visualization.rmd <==
---
title: "Data Wrangling with R, Chapter 3: Basic Data Visualization"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-06-29"
output: html_document
---

# Basic Data Visualization

## Load Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted) # An Alternative Conflict Resolution Strategy 
library(ggformula)  # Formula Interface to the Grammar of Graphics
library(patchwork)  # The Composer of Plots
library(tidyverse)  # Easily Install and Load the 'Tidyverse'
```

## Load Data

```{r}
# Dataset Load
data("mtcars")
```

## Univariate Visuals

### Histograms

```{r}
# Histogram of miles per gallon
mtcars %>%
  ggplot(aes(mpg)) +
  geom_histogram(
    breaks = seq(10, 35, by = 5),
    color  = "black", 
    fill   = "grey"
  ) +
  scale_y_continuous(
    breaks = seq(0, 12, by = 2),
    expand = c(0, 0)
  ) +
  labs(
    title = "Histogram of MPG",
    x     = "MPG",
    y     = "Frequency"
  ) +
  theme_classic()
```

### Customized histogram

```{r}
# Custom histogram
mtcars %>% 
  ggplot(aes(mpg)) +
  geom_histogram(
    breaks = seq(10, 35, by = 5),
    color  = "black",
    fill   = "royalblue"
  ) +
  scale_y_continuous(
    breaks = seq(0, 12, by = 2),
    expand = c(0, 0)
  ) +
  labs(
    title = "Histogram of Miles per Gallon (MPG)",
    x     = "MPG",
    y     = "Frequency"
  ) +
  theme_classic()
```

### Boxplots

```{r}
# Plot Boxplot of MPG
mtcars %>% 
  ggplot(aes(y = mpg)) +
  geom_boxplot(fill = "royalblue") +
  labs(
    title = "Boxplot of Miles per Gallon (MPG)",
    y     = "Miles per Gallon"
  ) +
  theme_classic() +
  theme(
    axis.text.x  = element_blank(),
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

Do cars with automatic transmission consume make less miles per gallon than the manual vehicles?

```{r}
# Boxplot AM vs MPG
mtcars %>% 
  ggplot(aes(as.factor(am), mpg)) +
  geom_boxplot(fill = "royalblue") +
  labs(
    title = "Types of Transmission and the Fuel Efficiency",
    x     = NULL,
    y     = "Miles per Gallon"
  ) +
  scale_x_discrete(
    labels = c("0" = "Automatic", "1" = "Manual")
  ) +
  theme_classic() +
  theme(
    axis.line.x     = element_blank(),
    axis.ticks.x    = element_blank(),
    legend.position = "none"
  )
```

#### Boxplot using formula notation

```{r}
# Boxplot with formula notation
mtcars %>% 
  gf_boxplot(
    mpg ~ as.factor(cyl),
    group = ~cyl,
    fill  = "royalblue",
  ) +
  labs(
    title = "Miles per Gallon by Number of Cylinders",
    x     = NULL,
    y     = "Miles per Gallon"
  ) +
  theme_classic() +
  theme(axis.ticks.x = element_blank())
```

### Density Plot

```{r}
# Density Plot
mtcars %>% 
  ggplot(aes(mpg)) +
  geom_density(fill = "red") +
  scale_x_continuous(
    limits = c(0, 50)
  ) +
  labs(
    title = "Density Plot of MPG",
    x     = NULL,
    y     = "Density"
  ) +
  theme_classic()
```

#### Compare two density plots

```{r}
mtcars %>% 
  ggplot(aes(mpg, group = am, fill = as.factor(am))) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(
    values = c("0" = "red", "1" = "blue"),
    labels = c("0" = "Automatic", "1" = "Manual")
  ) +
  scale_x_continuous(limits = c(0, 50)) +
  scale_y_continuous(breaks = seq(0, 0.10, by = 0.02)) +
  labs(
    fill  = NULL,
    title = "Density Plot of MPG",
    x     = NULL,
    y     = "Density"
  ) +
  theme_classic() +
  theme(
    legend.background = element_rect(color = "black"),
    legend.position   = c(0.88, 0.88),
    legend.title      = element_blank()
  )
```

## Two Variable Plots

### Scatterplots

```{r}
# Basic scatterplot
mtcars %>% 
  ggplot(aes(hp, mpg)) +
  geom_point(color = "royalblue", size = 2) +
  labs(
    title = "Effect of Horsepower on Miles per Gallon (MPG)",
    x     = "Horsepower",
    y     = "MPG"
  ) +
  theme_classic()
```

```{r}
# MPG vs Weight
mtcars %>%
  ggplot(aes(wt, mpg)) +
  geom_point(color = "royalblue", size = 5) +
  labs(
    title = "Weight vs. MPG",
    x     = "Weight",
    y     = "MPG"
  ) +
  theme_classic()
```

### Column Plot

```{r}
mtcars %>% 
  ggplot(aes(as.factor(cyl))) +
  geom_bar(fill = "royalblue") +
  labs(
    title = "Number of Cars by Cylinder",
    x     = NULL,
    y     = NULL
  ) +
  scale_y_continuous(
    breaks = c(seq(0, 14, by = 2)),
    expand = c(0, 0)
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
    )
```

### Bar Plot

```{r}
# Bar Plot
mtcars %>% 
  ggplot(aes(as.factor(vs))) +
  geom_bar(fill = "royalblue", color = "black") +
  scale_x_discrete(labels = c("0" = "V-shaped", "1" = "Straight")) +
  scale_y_continuous(
    expand = c(0, 0),
    limits = c(0, 20)
    ) +
  coord_flip() +
  labs(
    title = "Cars with V-Engine vs Straight",
    y     = NULL,
    x     = NULL
  ) +
  theme_classic() +
  theme(
    axis.line.y  = element_blank(),
    axis.ticks.y = element_blank()
  )
```

### Line Plot

```{r}
# Load dataset
data("AirPassengers")

# Column names
colnames <- c(
  "Jan",
  "Feb",
  "Mar",
  "Apr",
  "May",
  "Jun",
  "Jul",
  "Aug",
  "Sep",
  "Oct",
  "Nov",
  "Dec"
)

# Convert the data to a tibble
air_passengers_tbl <- matrix(
  AirPassengers,
  nrow  = 12,
  ncol  = 12,
  byrow = TRUE
) %>%
  as_tibble() %>%
  set_names(colnames) %>%
  mutate(year = 1949:1960) %>%
  pivot_longer(
    cols      = colnames,
    names_to  = "month",
    values_to = "passengers"
  ) %>%
  mutate(date = ymd(str_c(year, match(month, month.abb), "01", sep = "-")))

# Line Plot
air_passengers_tbl %>% 
  ggplot(aes(date, passengers)) +
  geom_line(color = "skyblue", linewidth = 1) +
  scale_x_date(
    date_breaks = "2 years",
    date_labels = "%Y" 
  ) +
  scale_y_continuous(breaks = seq(100, 600, by = 100)) +
  labs(
    title = "Evolution of the Number of Air Passengers from 1949-1960",
    x     = "Time",
    y     = "Passengers"
  ) +
  theme_classic()
```

If you want to plot multiple lines on the same figure using base R capabilities, the solution is to use the function lines() to add the extra series. 

```{r}
# Random X and Y
x  <- 1:12
y  <- rnorm(12) + 2
y2 <- rnorm(12) + 3
y3 <- rnorm(12) + 4

# Line Plot
tibble(x, y, y2, y3) %>%
  pivot_longer(
    cols      = c(y, y2, y3),
    names_to  = "line",
    values_to = "value"
  ) %>%
  ggplot(aes(x, value, color = line, linetype = line)) +
  geom_line(linewidth = 1) +
  scale_color_manual(values = c("royalblue", "coral", "darkgreen")) +
  scale_linetype_manual(values = c("solid", "twodash", "dashed")) +
  scale_x_continuous(breaks = seq(0, 12, by = 2)) +
  scale_y_continuous(breaks = seq(0, 6, by = 1)) +
  labs(
    color    = "Line",
    linetype = "Line",
    x        = "x",
    y        = "y"
  ) +
  theme_classic() +
  theme(
    legend.background = element_rect(color = "black"),
    legend.position   = c(0.96, 0.88),
    legend.title      = element_blank()
  )
```

## Plotting multiple variables

```{r, warning=FALSE}
# Scatterplot with 3 variables
mtcars %>%
  mutate(factor_cyl = as.factor(cyl)) %>%
  ggplot(aes(hp, mpg, color = factor_cyl, size = factor_cyl)) +
  geom_point(shape = 1) +
  scale_size_discrete(range = c(1, max(mtcars$cyl) / 2)) +
  labs(
    title = "MPG versus HP and Cylinder",
    x     = "HP",
    y     = "MPG"
  ) +
  theme_classic() +
  theme(
    legend.background = element_rect(color = "black"),
    legend.direction  = "horizontal",
    legend.position   = c(0.88, 0.88),
    legend.title      = element_blank()
  )
```

### Bar Plot with three variables (avoid!)

```{r}
# Stacked Bar Plot with Colors and Legend

# Prepare the data
mtcars_counts <- mtcars %>%
  count(cyl, vs, am) %>%
  mutate(across(where(~is_double(.)), as.factor))

# Create the stacked bar plot
p <- mtcars_counts %>%
  ggplot(aes(cyl, n, fill = vs)) +
  geom_col() +
  scale_fill_manual(
    values = c("1" = "coral", "0" = "royalblue"),
    labels = c("V", "Line")
  ) +
  labs(
    x        = "Cylinders",
    y        = "Count",
    title    = "Cars by Cylinder and Engine/Transmission",
    subtitle = "Engines with 8 cylinders tend to be constructed in V shape with automatic transmission"
  ) +
  theme_classic()

# Add lines for transmission types
am_counts <- mtcars_counts %>%
  summarize(
    n   = sum(n),
    .by = c(cyl, am)
  ) %>%
  # Add a dummy 'vs' column
  mutate(vs = "0") 

p + geom_line(
  data = am_counts,
  aes(cyl, n, group = am, color = am),
  linetype  = "solid",
  linewidth = 2
) +
  scale_color_manual(
    values = c("0" = "green", "1" = "yellow"),
    labels = c("Auto", "Manual"),
    name   = "Transmission"
  ) +
  theme(
    axis.line.x       = element_blank(),
    legend.background = element_rect(color = "black"),
    legend.box        = "horizontal",
    legend.position   = c(0.5, 0.8),
    legend.title      = element_blank()
  )
```

## Plot two graphics side by side

```{r, include=FALSE}
# Plot Graphics

# Create the scatter plot
p1 <- mtcars %>% 
  ggplot(aes(disp, hp)) +
  geom_point(color = "royalblue") +
  scale_y_continuous(breaks = seq(50, 300, by = 50)) +
  labs(
    title = "HP by Displacement",
    x     = "Displacement",
    y     = "HP"
    )

# Create the histogram
p2 <- ggplot(mtcars, aes(x = disp)) +
  geom_histogram(
    breaks = seq(50, 500, by = 50),
    fill   = "skyblue", 
    color  = "black") +
  scale_y_continuous(expand = c(0, 0)) +
  labs(
    title = "Histogram of Displacement",
    x     = "Displacement",
    y     = "Frequency"
    )

# Combine the plots using patchwork
combined_plot <- p1 + p2 & 
  theme_classic()
```

```{r}
# Print the combined plot
combined_plot
```
