==> 06_working_with_date_and_time_objects.rmd <==
---
title: "Data Wrangling with R, Chapter 6: Working with DateTime"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-07-14"
output: html_document
---

# Working with Date and Time Objects

## Load Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted) # An Alternative Conflict Resolution Strategy
  conflicts_prefer(dplyr::filter)
  conflicts_prefer(dplyr::lag)
library(hms)        # Pretty Time of Day
library(tidyverse)  # Easily Install and Load the 'Tidyverse'
```

## Creating a Date or DateTime objects

```{r}
# Checking today's system date
str_c("today(): ", today())

# Checking current system's time
str_c("now(): ", now())

# Creating a date object
str_c("as_date(0): ", as_date(0))

str_c("ymd(): ", ymd(20220714))

str_c("mdy(): ", mdy("Jul142022"))

str_c("yq(): ", yq("2010Q4"))

# Creating a time object
str_c("as_hms(): ", as_hms(43200))

str_c("ymd_hms(): ", ymd_hms(20220714150000))

# Creating a datetime object one year after 1970-01-1
#(60s*60min*24h*365d = 31,536,000 seconds)
str_c("as_datetime(): ", as_datetime(31536000))
```

### Datetime to Date

```{r}
# Datetime object
dt_tm <- ymd_hm("2022-01-02 03:04")

# Convert to just date object
as_date(dt_tm)
```

## Parsing DateTime Objects from Strings or Text

```{r}
# Creating a variable with 4 dates
dt <- c(
  "2000-01-01 01:01:01",
  "2005-02-02 02:02:02",
  "2010-03-03 03:03:03",
  "2020-04-04 04:04:04"
)
```

### Split Dates From a Variable

```{r}
# Assign it as datetime object
dt <- as_datetime(dt)

# years
year(dt)

# months
month(dt)

# weeks
week(dt)

#week day
wday(dt)

# days
day(dt)

# hours
hour(dt)

# minutes
minute(dt)

# seconds
second(dt)

# time zones
tz(dt)
```

### Gather DateTime Objects

```{r}
# separate variables
y_obj <- "2022"
m_obj <- "5"
d_obj <- "10"

# gather date
str_c(y_obj, m_obj, d_obj, sep = "-") %>% 
  ymd()
```

## Arithmetic Operations

### Simple Subtraction

```{r}
dt1 <- as_date("2022-06-01")
dt2 <- as_date("2022-05-01")

dt1 - dt2
```

### Period

```{r}
# Date
dt <- ymd("2000-01-01")

# Date minus a past date
today() - dt
```

```{r}
# Create a period to add or subtract
p <- years(x = 22) + months(x = 06) + days(x = 13)

# Another syntax
p <- period(c(22, 6, 13), c("year","month", "day"))

p
```

```{r}
# Operations
dt + p
dt - p
```

```{r}
# Date of the last maintenance
dt <- ymd("2021-01-15")

# Create a period to add or subtract
p <- years(x=1) + months(x=06) + days(x=1)

# Another syntax
p <- period( c(1, 6, 1), c("year","month", "day"))

# Calculation
cat(str_c("Next maintenance date is on: ", dt + p))
```

### Duration

```{r}
# Date
dt <- ymd("2000-01-01")

# Create a duration of 5 years
dw <- dyears(x = 5)

dw
```

```{r}
# Calculate warranty time 5 years after dt
warranty_end <- dt + dw

cat(str_c("Warranty ends on: ", warranty_end))
```

### Interval

```{r}
# Date
dt <- ymd("2022-01-01")

# Interval start
i <- interval(start = "2021-01-01", end = "2022-12-31")
i
```

```{r}
# Date within interval
dt %within% i
```

### Time Zone

```{r}
# System time zone
Sys.timezone()
```

```{r}
# Creating a datetime object in another timezone
ymd_hms("2022-01-01 00:00:00")
ymd_hms("2022-01-01 00:00:00", tz = "Europe/Paris")
```

### Display DateTime in Different Time Zone

```{r}
# Date creation
dt_dubai <- ymd_hms("2022-07-01 10:00:00", tz = "Asia/Dubai")
with_tz(dt_dubai, tzone = "America/New_York")
```

### Datetime in text

```{r}
# Lubridate parsing
mdy("The championship starts on 10/11-2000")
```

## Parsing with Regex

```{r}
# Text
t <- "The movie was launched on 10/10/1980. It was a great hype at that time, being the most watched movie on the weeks of 10/10/1980, 10/17/1980, 10/24/1980. Around ten years later, it was chosen as the best picture of the decade. The cast received the prize on 09/20/1990."

# Parse using regex
str_extract_all(t, "[0-9]+/[0-9]+/[0-9]+") %>%
  unlist() %>%
  mdy()
```

# Practice

The dataset to be used in this exercise is the Classic Rock, from fivethirtyeight.

To access the original dataset, got to: https://github.com/fivethirtyeight/data/tree/master/classic-rock

```{r}
# URL where the data is stored
url <- "https://raw.githubusercontent.com/fivethirtyeight/data/master/classic-rock/classic-rock-raw-data.csv"

# Load the dataset
tbl <- read_csv(url)

# Make a copy of the original data
tbl_original <- tbl

# View data frame in RStudio viewer
tbl
```

## Transform column time in datetime

```{r}
tbl_with_dates <- tbl %>% 
  mutate(TIME = as_datetime(TIME))

tbl_with_dates
```

## Select Variables of Interest

```{r}
tbl_selected <- tbl_with_dates %>% 
  select(
    "Song Clean",
    "ARTIST CLEAN",
    CALLSIGN,
    TIME,
    COMBINED,
    "First?"
  )

tbl_selected
```

## Add Extra Granularity for Time

```{r}
tbl_augmented <- tbl_selected %>%
  mutate(
    day     = day(TIME),
    weekday = wday(TIME),
    hour    = hour(TIME)
  )

tbl_augmented
```

## Visualizations

### Distribution of Songs Played Per Day of the Week

```{r}
tbl_unique <- tbl_augmented %>%
  distinct(
    `Song Clean`,
    .keep_all = TRUE
  )

song_by_weekday <- tbl_unique %>%
  count(weekday)

weekday <- song_by_weekday %>%
  pull(weekday)

song_by_weekday %>%
  ggplot(aes(as.factor(weekday), n)) +
  geom_col(fill = "royalblue", color = "black") +
  scale_x_discrete(
    labels = weekday %>% wday(label = TRUE)
  ) +
  scale_y_continuous(
    breaks = seq(0, 600, by = 100),
    expand = c(0, 0),
    limits = c(0, 600)
  ) +
  labs(
    title = "Number of unique songs by day of the week",
    x     = NULL,
    y     = "Distinct Songs"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

### Hour with the Most Songs Being Played

```{r}
song_by_hour <- tbl_augmented %>%
  count(hour)

song_by_hour %>%
  ggplot(aes(as.factor(hour), n)) +
  geom_col(fill = "royalblue", color = "black") +
  scale_x_discrete(breaks = seq(0, 23, by = 1)) +
  scale_y_continuous(
    breaks = seq(0, 8000, by = 2000),
    expand = c(0, 0),
    limits = c(0, 9000)) +
  labs(
    title = "Songs by hour",
    x     = "Hour",
    y     = "Songs Played"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

#### Filter and Investigate Songs Played at 23:00

```{r}
tbl_23 <- tbl_augmented %>% 
  filter(hour == 23) %>% 
  arrange(`Song Clean`, CALLSIGN, TIME)

tbl_23
```

#### Checking for Duplicates

```{r}
tbl_23 %>%
  distinct() %>%
  count() %>%
  mutate(difference = nrow(tbl_23) - n) %>%
  pull(difference)
```

#### Filter Only the Duplicated Rows

```{r}
dups <- tbl_23 %>%
  filter(duplicated(tbl_23))

dups
```

#### Test for Real Duplicates

Antti: This code checks if the next row number is not equal to 0 (which is the default value we provided for the lead function). If it is not equal to 0, the code calculates the number of equal columns between the current row and the next row. If it is equal to 0, the code assigns an NA value to the equal_cols variable for that row.

One thing to remember is that since we dropped the redundant year and month columns, the equal_cols == x needs to be 9 instead of 11 (as it was in the original)

```{r}
real_dup <- dups %>%
  mutate(
    row_number = row_number(),
    equal_cols = map2_dbl(
      row_number,
      lead(row_number, default = 0),
      ~if (.y != 0) {sum(dups[.x, ] == dups[.y, ])} 
      else NA
    )
  ) %>%
  filter(equal_cols == 9) %>%
  pull(row_number)

real_dup
```

#### Look at duplicates

```{r}

dups %>%
  filter(row_number() %in% c(9, 81))
```

#### Checking duplicates

```{r}
tbl_original %>%
  filter(
    TIME == ymd_hms("2014-06-19 23:54:25") & COMBINED %in% c(
      "Another One Bites the Dust by Queen",
      "Fly Like an Eagle by Steve Miller Band"
    )
  )
```

#### Remove duplicates

```{r}
tbl <- tbl_original %>%
  filter(!UNIQUE_ID %in% c("WCSX0949", "WCSX1021", "WCSX0955", "WCSX1005")) %>%

# Select variables and manipulate TIME variable
  select(
    song_clean   = "Song Clean",
    artist_clean = "ARTIST CLEAN",
    call_sign    = CALLSIGN,
    time         = TIME,
    combined     = COMBINED, 
    first        = "First?"
  ) %>%
  mutate(
    time    = as_datetime(time),
    day     = day(time),
    weekday = wday(time),
    hour    = hour(time)
  )

tbl
```

### Average Number of Songs Played by Hour

```{r}
# Group data by weekday
tbl_by_day <- tbl %>% 
  summarize(
    songs_ct = n(),
    .by = c(weekday, hour)
  ) %>% 

# Group the result by hour
  summarize(
    avg_songs_ct = mean(songs_ct),
    .by = hour
  )

# Line plot of Songs by weekday
tbl_by_day %>%
  ggplot(aes(hour, avg_songs_ct)) +
  geom_line(color = "royalblue", linewidth = 1.5) +
  scale_y_continuous(
    breaks = seq(0, 1200, by = 200),
    expand = c(0, 0)
  ) +
  labs(
    title = "Avg Number of Songs Played by Hour",
    x     = "Hour",
    y     = "Songs Played") +
  theme_classic()
```

### First Time Played by Hour

```{r}
# Filter only first time = 1
first_time <- tbl %>%
  filter(first == 1)

# First Time Songs by Hour
first_time <- first_time %>% 
  summarize(
    song_count = n(),
    .by        = hour
  )

# Line plot of First appearance Songs by hour
first_time %>%
  ggplot(aes(hour, song_count)) +
  geom_line(color = "royalblue", linewidth = 1.5) +
  scale_y_continuous(
    breaks = seq(0, 500, by = 100),
    expand = c(0, 0),
    limits = c(0, 500)
  ) +
  labs(
    title = "First Time Played by Hour",
    x     = "Hour",
    y     = "Songs Played") +
  theme_classic()
```

### Average Number of Songs Played by Radio by Day

```{r}
# Average by Radio by Day
by_radio <- tbl %>% 
  summarize(
    song_count = n(),
    .by = c(call_sign, weekday)
  ) %>% 
  summarize(
    avg_song = mean(song_count),
    .by = call_sign
  ) %>% 
  arrange(desc(avg_song)) %>%
  head(5)

# Bar Plot
by_radio %>% 
  mutate(call_sign = as.factor(call_sign)) %>%
  ggplot(aes(fct_reorder(call_sign, avg_song), avg_song)) +
  geom_col(fill = "royalblue", color = "black") +
  coord_flip() +
  scale_y_continuous(
    breaks = seq(0, 300, by = 50),
    expand = c(0, 0),
    limits = c(0, 320)
  ) +
  labs(
    title = "Average Number of Songs Played by Radio Station in a Day",
    x     = NULL,
    y     = NULL
  ) +
  theme_classic()
```
