==> 07_transformations_with_base_r.rmd <==
---
title: "Data Wrangling with R, Chapter 7: Transformations with Base R"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-07-21"
output: html_document
---

# Transformations with Base R

Antti: This whole time I've been transforming base R code to {tidyverse}. And I will continue to do so for the other chapters. This one is the exception, since the whole point of the chapter is using base R. I will still use things like the tibble and the pipe operator when it makes sense to do so.

## Load Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted) # An Alternative Conflict Resolution Strategy
library(data.table) # Extension of `data.frame`
library(skimr)      # Compact and Flexible Summaries of Data
library(tidyverse)  # Easily Install and Load the 'Tidyverse'
library(zoo)        # S3 Infrastructure for Regular and Irregular Time Series (Z's Ordered Observations)
```

## Dataset

The dataset to be used in this chapter is the _Adult Data Set_ from the UCI Machine Learning Repository.

__Dataset Credits:__
Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.

__URL Address__
https://archive.ics.uci.edu/ml/datasets/Adult

### Loading the Dataset

```{r}
# Load the datase
tbl <- read.csv(
  "https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.data",
  header      = FALSE,
  strip.white = TRUE
) %>%
  as_tibble()

# Add column names
colnames(tbl) <- c(
  "age",
  "workclass",
  "fnlwgt",
  "education",
  "education_num",
  "marital_status",
  "occupation",
  "relationship",
  "race",
  "sex",
  "capital_gain",
  "capital_loss",
  "hours_per_week",
  "native_country",
  "target"
)

tbl
```

## Slicing

```{r}
# Dataset dimensions
dim(tbl)
```

```{r}
# Slicing rows 1 to 3, all columns
tbl[1:3, ]
```

```{r}
# Slicing all rows, columns 1 to 3
tbl[, 1:3]

# Alternatively
tbl[, c("age", "workclass", "fnlwgt")]
```

```{r}
# Slicing with conditional, vector for cols 1 and 6
tbl[tbl$age > 30, c(1, 6)]

# Alternatively
tbl[tbl$age > 30, c("age", "marital_status")]
```

## Filtering

```{r}
# Filter
tbl %>%
  subset(age > 30)
```

```{r}
# Slice with subset, returning only the selected columns
tbl %>% 
  subset(age > 30, select = c(age, marital_status))
```

## Replace and Fill Values

### Replace

```{r}
# Check for NA values
sum(is.na(tbl))
```

```{r}
# Replace "?" using slicing
tbl$occupation[tbl$occupation == "?"]            <- NA
tbl[tbl$native_country == "?", "native_country"] <- NA

# Replace "?" with NA using gsub
tbl$workclass <- gsub(
  pattern     = "\\?",
  replacement = NA, 
  x           = tbl$workclass
)

# Replace "?" with ifelse
tbl$native_country <- ifelse(tbl$native_country == "?", NA, tbl$native_country)

# Check for NA values
sum(is.na(tbl))
```

### Check Missing Data by Column

```{r}
skim(tbl)
```

### Filling

```{r}
# Fill NAs with most frequent value
most_frequent <- names(table(tbl$workclass)[which.max(table(tbl$workclass))])
tbl$workclass[is.na(tbl$workclass)] <- most_frequent

# Occupation
most_frequent <- names(table(tbl$occupation)[which.max(table(tbl$occupation))])
tbl$occupation[is.na(tbl$occupation)] <- most_frequent

# Remove all NAs (the remaining in native_country)
tbl <- na.omit(tbl)

tbl
```

#### Filling NAs from Numeric Columns

##### With 0

```{r}
tbl_num <- tibble(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

tbl_num[is.na(tbl_num)] <- 0

tbl_num
```

##### With the Last Valid Number

```{r}
tbl_num <- tibble(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

for (i in 1:ncol(tbl_num)) {
  tbl_num[, i] <- na.locf(tbl_num[, i])
}

tbl_num
```

##### With the Next Valid Bumber

```{r}
df_num <- data.frame(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

for (i in 1:ncol(df_num)) {
  nas              <- is.na(df_num[, i])
  non_na_vals      <- na.locf(df_num[, i], fromLast = TRUE)
  df_num[, i][nas] <- non_na_vals[nas]
}

df_num
```

##### With Mean

```{r}
tbl_num <- tibble(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

tbl_num$A[is.na(tbl_num$A)] <- mean(tbl_num$A, na.rm = TRUE)

tbl_num
```

## Arranging Data

```{r}
# Arrange raw data
tbl[order(tbl$age, decreasing = TRUE), ]

# Arrange raw data using "-" instead of "decreasing = TRUE"
tbl[order(-tbl$age), ]
```

## Creating New Variables

```{r}
# Binarizing sex using ifelse()
tbl$sex_binary <- ifelse(tbl$sex == "Female", 1, 0)
tbl[ , c("sex", "sex_binary")]
```

## Encoding a Variable

```{r}
tbl$education_cd <- as.numeric(
  factor(
    tbl$education,
    levels = c(
      "Preschool",
      "1st-4th",
      "5th-6th",
      "7th-8th",
      "9th",
      "10th",
      "11th",
      "12th",
      "HS-grad",
      "Some-college",
      "Assoc-acdm",
      "Assoc-voc",
      "Bachelors",
      "Masters" ,
      "Doctorate"
    )
  )
)

tbl[, c("education", "education_cd")]
```

## Creating a Custom Calculation

```{r}
tbl$total_gain <- tbl$capital_gain - tbl$capital_loss
tbl$tax        <- ifelse(tbl$total_gain >= 15000, tbl$total_gain * 0.1, 0)

tbl[, c("capital_gain", "capital_loss", "education_cd", "total_gain")][order(-tbl$tax),]
```

## Binding

### Rows

```{r}
# Creating datasets A and B
A <- tbl[1:10, ]
B <- tbl[11:20, ]

# Append / bind rows
AB <- rbind(A, B)
AB
```

### Columns

```{r}
# Creating datasets A and B
A <- tbl[1:10, 1:3]
B <- tbl[1:10, c("education", "occupation")]

# Bind columns
AB <- cbind(A, B)
AB
```

## data.table

```{r}
# Load the dataset to RStudio
dt <- fread(
  "https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.data",
  header       = FALSE,
  strip.white  = TRUE,
  showProgress = FALSE
)

# Add column names
colnames(dt) <- c(
  "age",
  "workclass",
  "fnlwgt",
  "education",
  "education_num",
  "marital_status",
  "occupation",
  "relationship",
  "race",
  "sex",
  "capital_gain",
  "capital_loss",
  "hours_per_week",
  "native_country",
  "target"
)

dt
```

```{r}
# Syntax
  dt[dt$age > 50, .(age, occupation)]
```

### Group By and Summarize

```{r}
# Mean age Group By relationship
dt[, mean(age), by = relationship][order(-V1),]
```

```{r}
# Group By and count number of observations
dt[, .N, by = sex]
```

```{r}
# Group By and return mean of age by sex
dt[, mean(age), by = sex]
```

```{r}
# Group By and return sum of capital_gain by sex
dt[, sum(capital_gain), by = sex]
```

### Group and Visualize

```{r}
# Mean education years by education level
educ_yrs <- dt[, mean(education_num), by = education]
educ_yrs <- educ_yrs[order(V1)]

educ_yrs %>%
  as_tibble() %>%
  ggplot(aes(fct_reorder(education, V1), V1)) +
  geom_col(fill = "royalblue", color = "black") +
  coord_flip() +
  scale_y_continuous(
    expand = c(0, 0),
    breaks = seq(0, 20, by = 2)
  ) +
  labs(
    title = "Average Years of Education by Grade Level",
    x     = NULL,
    y     = NULL
  ) +
  theme_classic()
```

### Filling NAs from Numeric columns

#### With 0

```{r}
dt_num <- data.table(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

nafill(dt_num, type = "const", fill = 0)
```

#### With the Last Valid Number

```{r}
dt_num <- data.table(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

nafill(dt_num, type = "locf")
```

#### With the Next Valid Number

```{r}
dt_num <- data.table(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

nafill(dt_num, type = "nocb")
```

#### With Mean

```{r}
dt_num <- data.table(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

dt_num$A[is.na(dt_num$A)] <- mean(dt_num$A, na.rm = TRUE)

dt_num
```

### Arranging data

```{r}
# Group and order education years average by workclass
dt[, mean(education_num), by = workclass][order(-V1),]
```
