==> 08_transformations_with_tidyverse_libraries.rmd <==
---
title: "Data Wrangling with R, Chapter 8: Transformations with tidyverse"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-07-26"
output: html_document
---

# Transformations with Tidyverse Libraries

## Load Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted) # An Alternative Conflict Resolution Strategy
  conflicts_prefer(dplyr::filter)
library(tidyverse)  # Easily Install and Load the 'Tidyverse'
```

## Dataset

The dataset to be used in this chapter is the _Adult Data Set_ (also know as _Census Income_) from the UCI Machine Learning Repository.

__Dataset Credits:__
Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.

__URL Address__
https://archive.ics.uci.edu/ml/datasets/Adult

### Define column names

```{r}
header <- c(
  "age",
  "workclass",
  "fnlwgt",
  "education",
  "education_num",
  "marital_status",
  "occupation",
  "relationship",
  "race",
  "sex",
  "capital_gain",
  "capital_loss",
  "hours_per_week",
  "native_country",
  "target"
)
```

### Loading the dataset

```{r}
tbl <- read_csv(
  "https://archive.ics.uci.edu/ml/machine-learning-databases/adult/adult.data",
  col_names = header,
  trim_ws   = TRUE
)
```

```{r}
tbl
```

## Slicing

```{r}
# Slicing rows 1 to 5, columns 1 to 4.
tbl %>%
  slice_head(n = 5) %>% 
  select(age, workclass, fnlwgt, education)
```

```{r}
# Slicing with slice_min() and slice_max()
tbl %>% 
  slice_min(age, prop = 0.10)

tbl %>%
  slice_max(age, prop = 0.30)
```

```{r}
# Slice sample
tbl %>%
  slice_sample(
    n       = 10,
    replace = TRUE
  )
```

## Filtering

```{r}
# Filtering age over 30 years old
tbl %>%
  filter(age > 30)
```

```{r}
# Filter age > 30 and selecting age and marital_status
tbl %>% 
  filter(age > 30) %>% 
  select(marital_status, age)
```

```{r}
# Distinct - removing duplicates
tbl %>%
  distinct(sex)
```

## Group By and Summarize

```{r}
# Group by not summarized
tbl %>%
  group_by(workclass)
```

```{r}
# summarise without group_by() returning average age of the dataset
tbl %>%
  summarise(age_avg = mean(age))
```

```{r}
# Group by workclass and summarize mean age -> .by
tbl %>% 
  summarise(
    age_avg = mean(age),
    .by     = workclass
  )
```

```{r}
# Returns object grouped_tbl
workclass_groups <- tbl %>%
  group_by(workclass, sex) %>% 
  summarise(age_avg = mean(age))

workclass_groups
```

```{r}
# Returns object tibble
workclass_ungrouped <- tbl %>%
  summarise(
    age_avg = mean(age),
    .by     = c(workclass, sex)
  )

workclass_ungrouped
```

### Summary Functions

```{r}
# n() shows the count of observations in each group
tbl %>%
  summarise(
    n(),
    .by = workclass
  )

# Alternatively
tbl %>% 
  count(workclass, sort = TRUE)
```

```{r}
# n_distinct() shows the count of unique observations in each group
tbl %>%
  summarize(n_distinct(workclass))

tbl %>% 
  summarize(n_distinct(native_country), .by = workclass)
```

```{r}
# sum(!is.na()) shows the count of Non NA observations
tbl %>% summarise(sum(!is.na(workclass)))
```

```{r}
# first() shows the first age value in each group
# Similarly, you can use last() or nth()
tbl %>%
  summarise(
    first(age),
    .by = workclass
  )
```

```{r}
# quantile() shows the top number that meets the quantile percentage chosen
tbl %>% summarise(
  quantile(age, 0.5),
  .by = workclass
)
```

```{r}
# sd() shows standard deviation of a variable
tbl %>%
  summarise(
    sd(capital_gain),
    .by = workclass
  )
```

```{r}
# Across function, to apply the function mean to the selected columns
tbl %>%
  select(1, 3, 5, 11, 12, 13) %>%
  summarise(across(everything(), mean))
```

## Replace and Fill values

```{r}
na_by_column <- function(tbl) {
  tbl %>%
    summarise(across(everything(), ~sum(is.na(.x)))) %>%
    pivot_longer(
      cols      = everything(),
      names_to  = "Column",
      values_to = "NA count"
    ) %>%
    arrange(Column)
}
```

### Replace

```{r}
# Define a function to count '?' values in each column
count_question_marks <- function(x) {
  sum(x == "?", na.rm = TRUE)
}

count_question_marks(tbl)
```

```{r}
# Print the count of '?' values in each column
tbl %>%
  summarise(across(everything(), count_question_marks)) %>%
  pivot_longer(
    cols      = everything(),
    names_to  = "Column",
    values_to = "Count of '?'"
  ) %>%
  filter(`Count of '?'` > 0)
```

```{r}
# Replace '?' values with NA in a new dataset variable
tbl_replaced <- tbl %>%
  mutate(across(where(is.character), ~replace(., . == "?", NA)))
  
na_by_column(tbl_replaced)
```

### NA_IF

```{r}
# Replacing values "?" with NA (tidyverse 1.3.2)
tbl_replaced <- tbl %>%
  mutate(
    workclass      = na_if(workclass, "?"),
    occupation     = na_if(occupation, "?"),
    native_country = na_if(native_country, "?")
  )

na_by_column(tbl_replaced)
```

### Filling

```{r}
# Fill NA values with last or next valid value
tbl_replaced %>%
  fill(
    workclass,
    occupation,
    native_country,
    .direction = "down"
  )
```

```{r}
# Finding the most frequent entry in a column
get_most_frequent <- function(data, column) {
  data %>%
    summarise(
      freq = n(),
      .by  = all_of(column)
    ) %>%
    top_n(1, freq) %>%
    pull(column)
}

m_freq_workcls <- get_most_frequent(tbl, "workclass")
m_freq_occup   <- get_most_frequent(tbl, "occupation")

m_freq_workcls
m_freq_occup
```

```{r}
# Replace NA with the most frequent value
tbl_no_na <- tbl_replaced %>%
  mutate(
    workclass  = replace_na(workclass, m_freq_workcls),
    occupation = replace_na(occupation, m_freq_occup)
  ) %>%
  drop_na()

tbl_no_na
```

#### Filling NAs from Numeric Columns

```{r}
# Tibble
tbl_num <- tibble(
  A = c(1, 2, 2, 2, 3, NA),
  B = c(3, 4, 5, 3, NA, 0),
  C = c(1, 1, 1, NA, NA, 5)
)

# Fill NAs for numeric columns with statistics measurements
tbl_num %>% 
  replace_na(
    list(
      A = median(tbl_num$A, na.rm = TRUE),
      B = mean(tbl_num$B, na.rm = TRUE),
      C = min(tbl_num$C, na.rm = TRUE)
    )
  )
```

## Arranging Data

```{r}
# Arrange data in increasing order
tbl_no_na %>%
  arrange(native_country)
```

```{r}
# Arrange data in decreasing order
tbl_no_na %>%
  arrange(desc(native_country))
```

```{r}
# Group and order average net gain by education level
tbl_no_na %>%
  summarise(
    count        = n(),
    avg_net_gain = mean(capital_gain - capital_loss),
    .by          = education
  ) %>% 
  arrange(desc(avg_net_gain))
```

## Creating New Variables

Antti: Since the delimiter used in the original code was the word boundary, we couldn't replace the superseded _separate()_ function with _spread_wider_delim()_ and had to use _separate_wider_regex()_ instead.

```{r}
# Split variable target into sign and amount
tbl_no_na %>%
  separate_wider_regex(
    target,
    c(
      amt  = "\\D*(?<=\\b)",
      sign = ".*"
    )
  )
```

```{r}
# Unite variables sex, race and age
tbl_no_na %>% 
  unite(
    sex,
    race,
    age,
    col    = "description",
    sep    = "_",
    remove = FALSE
  )
```

## Creating a custom calculation

```{r}
# Create custom calculation: tax
tbl_no_na %>% 
  mutate(
    total_gain = capital_gain - capital_loss,
    tax        = if_else(total_gain >= 15000, total_gain * 0.1, 0)
  ) %>%
  arrange(desc(tax))
```

```{r}
# Change values from target to over or under
 tbl_no_na %>% 
  mutate(
    over_under = case_match(
      target,
      "<=50K" ~ "under",
      ">50K"  ~ "over"
    )
  ) %>%
  select(target, over_under)

# Alternatively
tbl_no_na %>% 
  mutate(over_under = if_else(target == "<=50K", "under", "over")) %>%
  select(target, over_under)
```

```{r}
# Determine if age is below or above the average
tbl_no_na %>% 
  mutate(
    age_avg            = mean(age, na.rm = TRUE),
    over_under_age_avg = if_else(age < age_avg, "Lower than avg", "Above the avg")
  ) %>%
  select(age, age_avg, over_under_age_avg)
```

## Reshape

### Wide to Long

```{r}
# set seed to generate the same random numbers every run
set.seed(42)

# Create a dataset
tbl_wide <- tibble(
  project = c("project1", "project2", "project3"),
  Jan = sample(1000:2000, 3),
  Feb = sample(1000:2000, 3),
  Mar = sample(1000:2000, 3)
)
  
# Convert from wide to long format
tbl_long <- tbl_wide %>% 
  pivot_longer(
    cols      = Jan:Mar,
    names_to  = "months",
    values_to = "expenses"
  )

tbl_long
```

### Long to Wide

```{r}
# Convert from long to wide format
tbl_long %>%
  pivot_wider(
    names_from  = "months",
    values_from = "expenses"
  )
```

# Joining Datasets

## Creating tables

```{r}
# Fact table
sales <- tibble(
  date       = c("2022-01-01", "2022-01-02", "2022-01-03", "2022-01-04", "2022-01-05"),
  store_cd   = c(1, 2, 3, 4, 5),
  product_cd = c(1, 2, 3, 4, 5),
  qty        = c(10, 12, 9, 12, 8), 
  sales      = c(30, 60, 45, 24, 32)
  )

sales

# Dimension store
stores <- tibble(
  store_cd   = c(1, 2, 3, 4, 6),
  address    = c("1 main st", "20 side st", "19 square blvd", "101 first st", "1002 retail ave"),
  city       = c("Main", "East", "West", "North", "South"), 
  open_hours = c("7-23", "7-23", "9-21", "9-21", "9-21")
  )

stores

# Dimension product
products <- tibble(
  product_cd   = c(1, 2, 3, 4, 6),
  description  = c("Soft drink", "Frozen snack", "Fruit", "Water", "Fruit 2"),
  unit_price   = c(3.0, 5.0, 5.0, 2.0, 4.0), 
  unit_measure = c("each", "each", "kg", "each", "kg")
  )

products
```

## Join tables using various types of joins

### Left Join

```{r}
sales %>% 
  left_join(products, by = "product_cd")
```

```{r}
# Left join with selected columns from products
sales %>%
  left_join(
    products %>%
      select(product_cd, description),
    by = "product_cd"
  )
```

### Right Join

```{r}
sales %>%
  select(store_cd, sales) %>% 
  right_join(stores, by = "store_cd")
```

### Inner Join

```{r}
sales %>% 
  inner_join(stores, by = "store_cd")
```

### Full Join

```{r}
sales %>%
  full_join(stores)
```

### Anti-Join

```{r}
sales %>%
  anti_join(products)
```

# Do more with __{tidyverse}__

## Load Data

```{r}
data("mtcars")
```

## Mapping Functions from {purrr}

### Map function

```{r}
# Apply a function to every element of a list or vector
mtcars %>% 
  select(hp, wt) %>% 
  map(mean)
```

### Map_int function

```{r}
# Apply a function to every element of a list or vector and return an integer vector
mtcars %>% 
  select(hp, wt) %>% 
  map_int(length)
```

## Binding Data

### Rows

```{r}
# Creating datasets A and B
A <- mtcars %>% 
  slice_head(n = 3)

B <- mtcars %>% 
  slice_tail(n = 3)

# Bind rows
AB <- A %>%
  bind_rows(B) %>% 
  as_tibble(rownames = "names")

AB
```

### Columns

```{r}
# Creating datasets A and B
A <- mtcars %>% 
  slice_head(n = 5) %>% 
  select(mpg:disp)

B <- mtcars %>% 
  slice_head(n = 5) %>% 
  select(hp:wt)

# Bind columns
AB <- A %>%
  bind_cols(B) %>% 
  as_tibble(rownames = "names")

AB
```

## Cumulative functions

```{r}
# Cumulative functions: calculate cumulative sum, product, mean, and maximum
mtcars %>%
  as_tibble(rownames = "names") %>%
  mutate(
    cumulative_sum_wt  = cumsum(wt),
    cumulative_prod_wt = cumprod(wt),
    cumulative_mean_wt = cummean(wt),
    cumulative_max_wt  = cummax(wt)
  ) %>%
  arrange(cumulative_sum_wt)
```

## Case When

```{r}
# Case_when function: apply multiple logical tests with corresponding results
mtcars %>%
  as_tibble(rownames = "names") %>%
  mutate(
    transmission_type = case_when(
      am == 0 ~ "automatic",
      am == 1 ~ "manual"
    )
  ) %>% 
  select(am, transmission_type)
```

## ggplot2

```{r}
# Basic scatterplot using ggplot2
mtcars %>%
  as_tibble(rownames = "names") %>%
  ggplot(aes(hp, mpg)) +
  geom_point(color = "royalblue", size = 4, alpha = 0.5) +
  labs(title = "Relationship between HP vs. MPG") +
  theme_classic()
```
