==> 09_exploratory_data_analysis.rmd <==
---
title: "Data Wrangling with R, Chapter 9: Exploratory Data Analysis"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-08-09"
output: html_document
---

# Exploratory Data Analysis

## Import Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted) # An Alternative Conflict Resolution Strategy
  conflicts_prefer(dplyr::filter)
library(GGally)     # Extension to "ggplot2'
library(scales)     # Scale Functions for Visualization     
library(skimr)      # Compact and Flexible Summaries of Data
library(statsr)     # Companion Software for the Coursera Statistics with R Specialization
library(tidyverse)  # Easily Install and Load the 'Tidyverse'
```

## Dataset

The dataset chosen for this project is from American Community Survey 2010-2012 Public Use Microdata Series and can be found in the raw form on the links below:

* Download data here: https://tinyurl.com/2ub3w436
* Documentation here: https://tinyurl.com/yckh444y

To download the data as seen in this project, go to fivethirtyeight web page on https://github.com/fivethirtyeight/data/tree/master/college-majors, where you will also find the data dictionary for the variables.

### Loading Data

```{r}
# Path
url <- "https://raw.githubusercontent.com/fivethirtyeight/data/master/college-majors/recent-grads.csv"

# Load dataset to RStudio
tbl <- read_csv(url)

# Keep a copy of our original data
tbl_original <- tbl
```
```{r}
tbl
```

```{r}
glimpse(tbl)
```

```{r}
nrow(tbl)
ncol(tbl)
dim(tbl)
```

### Transformations

```{r}
# Columns to change to factor
cols_to_factor <- c("Major", "Major_code", "Major_category")
```

```{r}
# Assign variables as factor
tbl <- tbl %>%
  mutate(across(all_of(cols_to_factor), factor))

# Check result
glimpse(tbl)
```

```{r}
# Remove scientific notation
options(scipen = 999, digits = 3)

# Coefficient of Variance
stats <- skim(tbl)

stats$numeric.sd / stats$numeric.mean
```

```{r}
skim(tbl)
```

```{r}
summary(tbl)
```

### Missing Data

```{r}
# Drop NA
tbl_clean <- tbl %>%
  drop_na()

# New Dimensions
dim(tbl_clean)
```

## Exploring and Visualizing Data

### Univariate Analysis

#### Histogram

```{r}
# Select numeric columns and convert to long format
numeric_data <- tbl_clean %>%
  select(where(is.numeric)) %>%
  pivot_longer(
    cols      = everything(),
    names_to  = "variable",
    values_to = "value"
  )

# Function to create histograms for each variable
create_histogram <- function(variable, data) {
  p <- data %>%
    filter(variable == !!variable) %>%
    ggplot(aes(value)) +
    geom_histogram(
      bins  = 20,
      fill  = "royalblue",
      color = "gray"
    ) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    labs(
      x        = "Value",
      title    = str_c("Histogram of ", variable)
    ) +
    theme_classic()
  
  return(p)
}

# Create a list of ggplot objects using purrr::map()
plots <- map(
  unique(
    numeric_data %>% 
      pull(variable)
  ),
  create_histogram,
  data = numeric_data
)

# To view a specific plot, use the index, e.g., plots[[1]]...
plots
```

#### Boxplot

```{r}
# Select numeric columns and convert to long format
numeric_data <- tbl_clean %>%
  select(where(is.numeric)) %>%
  pivot_longer(
    cols      = everything(),
    names_to  = "variable",
    values_to = "value"
  )

# Function to create boxplots for each variable
create_boxplot <- function(variable, data) {
  p <- data %>%
    filter(variable == !!variable) %>%
    ggplot(aes(y = value)) +
    geom_boxplot(fill = "royalblue", color = "gray") +
    labs(
      y     = NULL,
      title = paste("Boxplot of", variable)
    ) +
    theme_classic()
  
  return(p)
}

# Create a list of ggplot objects using purrr::map()
boxplots <- map(
  unique(
    numeric_data$variable
  ),
  create_boxplot,
  data = numeric_data
)

# To view a specific plot, use the index, e.g., boxplots[[1]]
boxplots
```

#### QQ-plot

```{r}
# Select numeric columns and convert to long format
numeric_data <- tbl_clean %>%
  select(where(is.numeric)) %>%
  pivot_longer(
    cols      = everything(),
    names_to  = "variable",
    values_to = "value"
  )

# Function to create QQ-plots for each variable
create_qq_plot <- function(variable, data) {
  p <- data %>%
    filter(variable == !!variable) %>%
    ggplot(aes(sample = value)) +
    stat_qq() +
    stat_qq_line(color = "royalblue", linewidth = 1) +
    labs(
      x     = NULL,
      y     = "Theoretical",
      title = str_c("QQ-plot of ", variable)
    ) +
    theme_classic()
  
  return(p)
}

# Create a list of ggplot objects using purrr::map()
qq_plots <- map(
  unique(numeric_data$variable),
  create_qq_plot,
  data = numeric_data
)

# To view a specific plot, use the index, e.g., qq_plots[[1]]
qq_plots
```

#### Normality Test for Unemployment_rate

```{r}
shapiro.test(tbl_clean %>% pull(Unemployment_rate))
```

### Multivariate Analysis

```{r}
# Check linear relationship and correlations
tbl_clean %>%
  select(-c(Rank, Major_code, Major, Major_category)) %>%
  ggpairs(progress = FALSE)
```

```{r}
# Plot correlation matrix as a heatmap
correlation_matrix <- tbl_clean %>%
  select(-c(Rank, Major_code, Major, Major_category)) %>%
  cor() %>%
  as.data.frame() %>%
  rownames_to_column(var = "Var1") %>%
  pivot_longer(
    cols      = -Var1,
    names_to  = "Var2",
    values_to = "Correlation"
  ) %>% as_tibble()

correlation_matrix %>%
  ggplot(aes(Var1, Var2, fill = Correlation)) +
  geom_tile(color = "black") +
  geom_text(aes(label = round(Correlation, 2)), color = "black") +
  scale_fill_gradient2(
    low      = "red",
    high     = "blue",
    mid      = "white",
    midpoint = 0,
    limits   = c(-1, 1)
  ) +
  labs(
    fill = "Correlation",
    x    = NULL,
    y    = NULL
  ) +
  theme_classic() +
  theme(
    axis.text.x     = element_text(angle = 90),
    legend.position = "none"
    )
```

#### Top 10 Majors?

```{r}
# Select only top 10 for plotting
top10_low_unemployment <- tbl_clean %>% 
  select(Major, Unemployment_rate) %>% 
  arrange(Unemployment_rate) %>% 
  head(10)

# Plot
top10_low_unemployment %>%
  ggplot(
    aes(
      Unemployment_rate,
      fct_reorder(Major %>% str_to_title(), Unemployment_rate, .desc = TRUE)
    )
  ) +
  geom_col(color = "royalblue", fill = "royalblue") +
  geom_text(
    aes(label = round(Unemployment_rate, 3)),
    size  = 3,
    hjust = 1.2,
    color = "white"
    ) +
  scale_x_continuous(expand = c(0, 0)) +
  labs(
    title = "Top 10 Majors with the Lowest Unemployment Rate",
    x     = NULL,
    y     = NULL
  ) +
  theme_classic() +
  theme(
    axis.text.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

```{r}
# Select the top 10 majors with the highest proportion and lowest unemployment rate
top10_proportional <- tbl_clean %>% 
  mutate(proportion = Total / sum(Total)) %>%
  select(Major, Unemployment_rate, proportion) %>%
  arrange(desc(proportion), Unemployment_rate) %>% 
  head(10)

# Plot the top 10 majors with the highest proportion and lowest unemployment rate
top10_proportional %>%
  ggplot(
    aes(
      Unemployment_rate,
      fct_reorder(Major %>% str_to_title(), Unemployment_rate, .desc = TRUE)
    )
  ) +
  geom_col(color = "royalblue", fill = "royalblue") +
  geom_text(
    aes(label = round(Unemployment_rate, 3)),
    size  = 3,
    hjust = 1.2,
    color = "white"
  ) +
  scale_x_continuous(expand = c(0, 0)) +
  labs(
    title = "Majors with Lowest Unemployment Rate [Normalized]",
    x = NULL,
    y = NULL
  ) +
  theme_classic() +
  theme(
    axis.text.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

#### More Specialized Majors

```{r}
# Difference collge - non-college jobs
top_college_jobs <- tbl_clean %>% 
  mutate(College_jobs_pct = College_jobs / (College_jobs + Non_college_jobs)) %>%
  select(Major_category, Major, College_jobs_pct, Median) %>%
  arrange(desc(College_jobs_pct))
```

```{r}
# Look at the top 10
top_10_college_jobs <- top_college_jobs %>%
  head(10)

top_10_college_jobs
```

```{r}
# Look at the bottom 10 
bottom_10_college_jobs <- top_college_jobs %>%
  tail(10)

bottom_10_college_jobs
```

```{r}
# Calculate the mean college jobs percentage by major category
mean_college_jobs_pct <- top_college_jobs %>%
  summarise(
    mean_college_pct = mean(College_jobs_pct),
    med_salary       = mean(Median),
    .by              = Major_category
  ) %>% 
  arrange(desc(mean_college_pct))

mean_college_jobs_pct
```

#### Best Median Value Paying Jobs

```{r}
# Select the top 10 majors with the highest median salary
top_10_median_salary <- tbl_clean %>%
  arrange(desc(Median)) %>%
  head(10)

top_10_median_salary %>% 
  ggplot(aes(Median, fct_reorder(Major %>% str_to_title(), Median))) + 
  geom_col(fill="royalblue") +
  geom_text(
    aes(label = label_dollar()(Median)),
    size  = 3,
    hjust = 1.2,
    color = "white"
  ) +
  scale_x_continuous(expand = c(0, 0)) +
  labs(
    title = "Top 10 Majors with the Highest Median Salary",
    x     = NULL,
    y     = NULL, 
    ) +
  theme_classic() +
  theme(
    axis.text.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

```{r}
# Add variable median salaries
tbl_clean_with_median_salaries <- top_college_jobs %>%
  mutate(
    median_pay = median(Median),
    .by        = Major_category
  )

# Boxplots of the median payments by major
tbl_clean_with_median_salaries %>% 
  ggplot(aes(fct_reorder(Major_category, median_pay), Median)) + 
  geom_boxplot(fill="royalblue") +
  scale_y_continuous(labels = label_dollar()) +
  expand_limits(
    x = c(0, NA),
    y = c(0, NA)
  ) +
  coord_flip() +
  labs(
    title = "Median by Major Category",
    x     = NULL,
    y     = NULL
    ) +
  theme_classic()
```

#### What's Gender Got to Do with It?

```{r}
# Calculate the correlation between the share of women enrolled and the unemployment rate
correlation_women_unemployment <- cor(
  tbl_clean %>% pull(ShareWomen),
  tbl_clean %>% pull(Unemployment_rate),
  method = "spearman"
)

correlation_women_unemployment
```

```{r}
# Plot the share of women enrolled vs. unemployment rate
tbl_clean %>% 
  ggplot(aes(ShareWomen, Unemployment_rate)) +
  geom_point(color="royalblue") +
  labs(
    title    = "Share of Women Enrolled vs. Unemployment rate",
    subtitle = str_glue(
      "Correlation: {round(correlation_women_unemployment, 3)} (Weak correlation)"),
    x        = "Share of Women",
    y        = "Unemployment Rate"
  ) +
  theme_classic()
```

```{r}
# Calculate the correlation between the share of women enrolled and low-wage jobs
correlation_women_low_wage <- cor(
  tbl_clean %>% pull(ShareWomen),
  tbl_clean %>% pull(Low_wage_jobs)
)
```

```{r}
# Plot the share of women enrolled vs. low-wage jobs
tbl_clean %>%
  ggplot(aes(ShareWomen, Low_wage_jobs)) + 
  geom_point(color = "royalblue") +
  labs(
    title    = "Share of Women Enrolled vs. Low-Wage Jobs",
    subtitle = str_glue(
      "Correlation: {round(correlation_women_low_wage, 3)} (Weak correlation)"),
    x        = NULL,
    y        = NULL
  ) +
  theme_classic()
```

```{r}
# Create new column for Share Women higher or lower than 50%
tbl_clean_women_shares <- tbl_clean %>% 
  mutate(
    over_under = case_when(
      ShareWomen > 0.5 ~ "higher",
      TRUE             ~ "lower"
    )
  )

tbl_clean_women_shares
```

```{r}
set.seed(42)

# Perform hypothesis test to compare the average low-wage jobs for majors with more than 50% share of women and majors with less than 50% share of women
higher_women <- tbl_clean_women_shares %>%
  filter(over_under == "higher") %>%
  select(Major, Major_category, Low_wage_jobs)

lower_women <- tbl_clean_women_shares %>%
  filter(over_under == "lower") %>%
  select(Major, Major_category, Low_wage_jobs)

# Shapiro-Wilk's test for normality
shapiro_test_higher <- shapiro.test(higher_women$Low_wage_jobs)
shapiro_test_lower  <- shapiro.test(lower_women$Low_wage_jobs)

shapiro_test_higher
shapiro_test_lower
```

```{r}
set.seed(42)

# Sampling from Higher Share Women
higher_women_n <- higher_women %>%
  rep_sample_n(
    size    = 1000,
    reps    = 100,
    replace = TRUE
  ) %>%
  summarise(mu = mean(Low_wage_jobs))

# Sampling from Lower Share Women
lower_women_n <- lower_women %>%
  rep_sample_n(
    size    = 1000,
    reps    = 100,
    replace = TRUE
  ) %>%
  summarise(mu = mean(Low_wage_jobs))

# Normality tests
shapiro.test(higher_women_n %>% pull(mu))
shapiro.test(lower_women_n %>% pull(mu))
```

```{r}
# T-test to check if the averages of both groups
t.test(
  x           = higher_women_n %>% pull(mu),
  y           = lower_women_n %>% pull(mu),
  alternative = "two.sided"
)
```

```{r}
# Group data
median_salary_w <- tbl_clean_women_shares %>% 
  summarise(
    med_sal = mean(Median),
    .by     = over_under
    )

# Plot a Boxplot graphic
tbl_clean_women_shares %>% 
  ggplot(aes(over_under, Median)) + 
  geom_boxplot(fill = "royalblue") +
  scale_y_continuous(labels = label_dollar()) +
  labs(
    title    = "Average Salary When the Share of Women is lower/higher than 50%",
    subtitle = str_glue(
      "The average salary for majors with more women enrolled is lower than the majors with less women,
      reinforcing the perception that women are getting lower salaries."),
    x        = "50% Share of Women",
    y        = NULL
  ) +
  geom_text(
    data = median_salary_w,
    aes(
      over_under,
      med_sal,
      label = label_dollar(accuracy = 1)(med_sal)
    ),
    size  = 4,
    vjust = 1,
    color = "white"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

```{r}
# Calculate the percentage difference between averages
percentage_difference <- median_salary_w %>%
  summarise(
    difference = (last(med_sal) - first(med_sal)) / last(med_sal)
  ) %>%
  pull(difference) %>%
  round(3) * 100

# Print the percentage difference
str_glue("The percentual difference between both averages is {percentage_difference} %")
```

#### Correlation Between College Jobs and Unemployment Rate?

```{r}
# Correlation between ShareWomen vs Unemployment_rate
correlation_collegejobs_unemployment <- cor(
  tbl_clean %>% pull(College_jobs),
  tbl_clean %>% pull(Unemployment_rate),
  method = "spearman"
)

correlation_collegejobs_unemployment
```

```{r}
tbl_clean %>%
  ggplot(aes(College_jobs, Unemployment_rate)) +
  geom_point(color="royalblue") +
  labs(
    title    = "College Jobs vs. Unemployment rate",
    subtitle = str_glue(
      "Correlation: {round(correlation_collegejobs_unemployment, 3)} (Weak correlation)"
    ),
    x = NULL,
    y = NULL
  ) +
  theme_classic()
```
