==> 10_introduction_to_ggplot2.rmd <==
---
title: "Data Wrangling with R, Chapter 10: Intro to ggplot2"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-08-24"
output: html_document
---

# Intro to {ggplot2}

## Load Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted) # An Alternative Conflict Resolution Strategy
library(patchwork)  # The Composer of Plots
library(tidyverse)  # Easily Install and Load the 'Tidyverse'
```

## My First {ggplot2} Graphic

```{r}
# Set a seed for reproduce the same results
set.seed(42)

# Create a sample dataset
tbl <- tibble(
  var1 = runif(20),
  var2 = rnorm(20)
)

tbl %>%
  ggplot(aes(var1, var2)) +
  geom_point() +
  labs(title = "My first ggplot2 graphic") +
  theme_classic()
```

## Dataset

```{r}
data("mtcars")

cars_tbl <- mtcars %>% 
  as_tibble(rownames = "names")

cars_tbl
```

## Plot Types

### Histogram

```{r}
cars_tbl %>% 
  ggplot(aes(mpg)) +
  geom_histogram( 
    bins  = 20,
    color = "lightgray",
    fill  = "royalblue"
  ) +
  scale_y_continuous(expand = c(0, 0)) +
  labs(title = "Histogram of Miles per Gallon") +
  theme_classic()
```

### Boxplot

```{r}
# Boxplot of MPG
cars_tbl %>%
  ggplot(aes(y = mpg)) + 
  geom_boxplot(fill = "royalblue") + 
  labs(title = "Miles per Gallon (MPG)") +
  theme_classic()
```

```{r}
# Boxplot of MPG by engine type
cars_tbl %>% 
  ggplot(aes(factor(vs), mpg)) + 
  geom_boxplot(fill = "royalblue") +
  scale_x_discrete(labels = c("1" = "V-shaped", "0" = "Line-shaped")) +
  labs(
    x     = "Engine shape",
    title = "A comparison between V-shaped vs. Line-shaped engines and the effect on MPG"
  ) +
  theme_classic()
```

### Scatterplot or Points Plot

```{r}
# Scatterplot weight versus mpg
cars_tbl %>% 
  ggplot(aes(wt, mpg)) +
  geom_point(
    color = "royalblue",
    size  = 4,
    shape = 15,
    alpha = 0.7
  ) +
  labs(
    x        = "Weight of the cars",
    y        = "Miles per gallon",
    title    = "How does the weight affect MPG in cars?",
    subtitle = "As the weight increases, the car will make less miles per gallon"
  ) +
  theme_classic()
```

### Bar Plot

#### Single Variable

```{r}
cars_tbl %>%
  ggplot(aes(factor(am))) + 
  geom_bar(fill = "royalblue") +
  scale_x_discrete(labels = c("1" = "Manual", "0" = "Automatic")) +
  scale_y_continuous(
    expand = c(0, 0),
    breaks = seq(0, 20, by = 5),
    limits = c(0, 20)
  ) +
  labs(
    x     = NULL,
    y     = NULL,
    title = "Count of observations by transmission type"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

#### Multiple Variables

```{r}
# Bar plot with two variables
cars_tbl %>% 
  ggplot(aes(factor(am), mpg)) + 
  # statistic calculation - mean value
  geom_bar(
    stat = "summary",
    fun  = "mean",
    fill = "royalblue"
  ) +
  scale_x_discrete(labels = c("1" = "Manual", "0" = "Automatic")) +
  scale_y_continuous(
    expand = c(0, 0),
    breaks = seq(0, 25, by = 5)
  ) +
  labs(
    x     = NULL,
    y     = NULL,
    title = "Average MPG by transmission type"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

#### With stat_summary()

```{r}
# Another syntax
cars_tbl %>% 
  ggplot(aes(factor(am), mpg)) + 
  # statistic calculation - mean value
  stat_summary(
    fun  = "mean",
    geom = "bar",
    fill = "royalblue"
  ) +
  scale_x_discrete(labels = c("1" = "Manual", "0" = "Automatic")) +
  scale_y_continuous(
    expand = c(0, 0),
    breaks = seq(0, 25, by = 5)
  ) +
  labs(
    x     = NULL,
    y     = NULL,
    title = "Averate MPG by transmission type"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

### Column plot

```{r}
# Column plot for MPG by transmission type
cars_tbl %>% 
  ggplot(aes(factor(am), mpg)) +
  geom_col(fill = "royalblue") +
  scale_x_discrete(labels = c("1" = "Manual", "0" = "Automatic")) +
  scale_y_continuous(
    expand = c(0, 0),
    breaks = seq(0, 350, by = 50),
    limits = c(0, 350)
  ) +
  labs(
    x     = NULL,
    y     = NULL,
    title = "MPG by transmission type"
  ) +
  theme_classic() +
  theme(
    axis.line.x  = element_blank(),
    axis.ticks.x = element_blank()
  )
```

#### Stacked or Side-by-Side

```{r}
# Base plot
p <- cars_tbl %>% 
  ggplot(aes(factor(cyl), fill = factor(vs))) +
  scale_fill_discrete(labels = c("0" = "Line-shaped", "1" = "V-shaped")) +
  scale_y_continuous(expand = c(0, 0)) +
  labs(x = "Cylinders")

# Bar plot stacked
g1 <- p +
  geom_bar(position = "stack")

# Bar plot side
g2 <- p +
  geom_bar(position = "dodge")

# Bar plot fill
g3 <- p +
  geom_bar(position = "fill")

# Patchwork to put three plots on the same line
((g1| g2| g3) +
  plot_layout(guides = "collect") & theme(legend.position  = "bottom")) *
  labs(
    y = NULL
  ) *
  theme_classic() *
  theme(
    legend.direction = "horizontal",
    legend.title     = element_blank(),
    axis.line.x      = element_blank(),
    axis.ticks.x     = element_blank()
  )
```

### Line plot

#### Single Line

```{r}
# Seed for reproduction
set.seed(42)

# Car sales
sales <- tibble(
  month  = factor(1:12),
  sales  = as.integer(abs(rnorm(12) * 10000)),
  sales2 = as.integer(abs(rnorm(12) * 11000))
)

sales
```

```{r}
# Simple Line plot
sales %>%
  ggplot(aes(month, sales, group = 1)) +
  geom_line(linewidth = 1, color = "darkgreen") +
  labs("Car sales throughout the months") +
  theme_classic()
```

#### Multiple Lines

```{r}
# Line plot with two variables
sales %>% 
  ggplot() +
  geom_line(aes(month, sales, group = 1, color = "sales year 1"), linewidth = 1, linetype = 2) +
  geom_line(aes(month, sales2, group = 1, color = "sales year 2"), linewidth = 1, linetype = 1) +
  labs(title = "Car sales throughout the months - Two year comparison") +
  theme_classic() +
  theme(
    legend.title = element_blank()
  )
```
#### Multiple Lines with pivot_longer()

```{r}
# Alternatively using pivot_longer first
sales %>%
  pivot_longer(cols = sales:sales2, names_to = "group", values_to = "values") %>%
  ggplot(aes(month, values, group = group, color = group)) +
  geom_line(linewidth = 1) +
  scale_color_discrete(labels = c("sales" = "1", "sales2" = "2")) +
  labs(
    title  = "Car sales throughout the months - Two year comparison",
    color  = "Sales year") +
  theme_classic()
```

#### Smooth Geometry

```{r}
# Smooth line
cars_tbl %>%
  ggplot(aes(hp, mpg)) +
  geom_point(aes(color= factor(vs))) +
  geom_smooth(method = "loess") +
  scale_color_discrete(labels = c("0" = "Line-shaped", "1" = "V-shaped")) +
  labs(
    color = "Engine Shape",
    x     = "HP",
    y     = "MPG"
    ) +
  theme_classic()
```

## Themes

```{r}
# Base Plot
p <- cars_tbl %>% 
  ggplot(aes(factor(am))) +
  geom_bar(fill= "royalblue") +
  labs(x = "am")

# Theme BW
g1 <- p +
  labs(title = "BW") +
  theme_bw()

# Theme Classic
g2 <- p +
  labs(title = "Classic") +
  theme_classic()

# Theme Light
g3 <- p +
  labs(title = "Light") +
  theme_light()

# Theme Dark
g4 <- p +
  labs(title = "Dark") +
  theme_dark()

# Theme Gray
g5 <- p +
  labs(title = "Gray") +
  theme_gray()

# Theme Linedraw
g6 <- p +
  labs(title = "Linedraw") +
  theme_linedraw()

# Theme Minimal
g7 <- p +
  labs(title = "Minimal")+
  theme_minimal()

# Theme void
g8 <- p +
  labs(title = "Void")+
  theme_void()

(g1 | g2 | g3 | g4) / (g5 | g6 | g7 | g8)
```
