==> 12_other_data_visualization_options.rmd <==
---
title: "Data Wrangling with R, Chapter 12: Other Data Visualization Options"
author: "Gustavo R. Santos (original) | Antti Rask (modifications)"
date: "2022-08-29"
output: html_document
---

# Other Data Visualization Options
 
## Load Libraries

```{r, message=FALSE, warning=FALSE}
library(conflicted)  # An Alternative Conflict Resolution Strategy
  conflicts_prefer(dplyr::filter)
library(ggwordcloud) # A Word Cloud Geom for 'ggplot2' 
library(officer)     # Manipulation of Microsoft Word and PowerPoint Documents
library(tidytext)    # Text Mining using 'dplyr', 'ggplot2', and Other Tidy Tools
library(tidyverse)   # Easily Install and Load the 'Tidyverse'
```

## Plotting Graphic in Microsoft Power BI

```{r}
# Data frame created
data <- tibble(
  dist1 = rnorm(100),
  dist2 = runif(100)
)

# Save to disk
write_csv(data, "data/example_data.csv")
```

```{r}
# To plot this graphic in Power BI, copy and paste this code in the tool
data %>% 
  ggplot() +
  geom_histogram(
    aes(dist1),
    bins  = 10,
    color = "white",
    fill  = "royalblue"
  ) +
  scale_y_continuous(expand = c(0, 0)) +
  theme_classic()
```
```{r}
dataset <- tibble(dist1 = data %>% pull(dist1))

head(dataset)
```

## Word Cloud

```{r}
# Read word document and extract textual information
text <- read_docx("data/Data_Wrangling_With_R_Chapter10_for_wordcloud.docx") %>% 
  docx_summary() %>%
  as_tibble() %>% 
  filter(text != "") %>%
  pull(text) %>% 
  str_squish() %>% 
  as_tibble()

text
```

```{r}
# Tokenize - one word is one token
clean_tokens <- text %>%
  unnest_tokens(output = word, input = value) %>%
  filter(str_detect(word, "\\D")) %>%
  filter(!str_detect(word, "[:punct:]")) %>%
  anti_join(stop_words) %>%
  count(word, sort = TRUE)

clean_tokens
```

```{r, warning=FALSE}
clean_tokens %>%
  mutate(
    angle = 90 * sample(
      c(0, 1),
      n(),
      replace = TRUE,
      prob = c(60, 40)
    )
  ) %>% 
  ggplot(
    aes(
      label = word,
      size  = n,
      angle = angle,
      color = factor(sample.int(10, nrow(clean_tokens), replace = TRUE))
    )
  ) +
  geom_text_wordcloud(
    shape      = "circle",
    area_corr  = TRUE,
    rm_outside = TRUE
  ) +
  scale_size_area(max_size = 30) +
  theme_minimal()
```
